+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ref(nextjs): Clean up minor things (#258)
 - ref(nextjs): Replace old Next.js wizard (#262)
 - fix(rn): Show loader when installing dependencies (#264)
+- feat(rn): Add code snippet to send the first Sentry Error ([#263](https://github.com/getsentry/sentry-wizard/pull/263))
 
 ## 3.1.0